Move menu content
