+++ web/client/src/components/navigation/menu/soro-navigation-menu-content.vue
